requirements.txt
flask
playwright==1.49.0
pillow
pytest